Improve performance of computing job runtime stats
Summary: We already have the functionality to return the count of tasks per state, so using that will significantly lower the preasure on the DB, when computing job runtime stats.

Reviewers: min, mu, slu, fgalilee, kejlberg, #peloton, avyas

Reviewed By: #peloton, avyas

Subscribers: jenkins, avyas

Differential Revision: https://code.uberinternal.com/D1171339

+++ storage/cassandra/store_test.go
@@ -774,7 +774,7 @@ func (suite *CassandraStoreTestSuite) TestGetTaskStateSummary() {
 	suite.Nil(err)
 	suite.Equal(len(taskStateSummary), len(task.TaskState_name))
 	for _, state := range task.TaskState_name {
-		suite.Equal(taskStateSummary[state], 2)
+		suite.Equal(taskStateSummary[state], uint32(2))
 	}
 }
